feat:style and code refactor details page
