+++ src/components/details/post-add-comment/index.module.css
@@ -1,0 +1,16 @@
+.container {
+    border-top: 1px solid rgb(207, 207, 207);
+    padding-top: 15px;
+    display: flex;
+}
+
+.textarea {
+    width: 100%;
+    border: 1px solid transparent;
+    outline: none;
+    resize: none;
+}
+
+.textarea::placeholder {
+    color: rgb(185, 185, 185);
+}
