+++ src/components/details/post-header/index.module.css
@@ -1,0 +1,42 @@
+.header {
+    display: flex;
+    justify-content: space-between;
+}
+
+.profile {
+    display: flex;
+    align-items: center;
+}
+
+.avatar {
+    margin-right: 15px;
+    border-radius: 50px;
+    width: 35px;
+    height: 35px;
+    object-fit: cover;
+    cursor: pointer;
+}
+
+.username {
+    color: black;
+    font-weight: bold;
+    font-size: 14px;
+}
+
+.location {
+    font-size: 12px;
+    color: gray;
+}
+
+.followBtn {
+    background-color: #0095f6;
+    border: 1px solid #dbdbdb;
+    border: 1px solid rgba(var(--ca6, 219, 219, 219), 1);
+    color: #fff;
+    border-radius: 4px;
+    position: relative;
+    text-decoration: none;
+    display: block;
+    font-weight: 600;
+    padding: 5px 9px;
+}
