+++ src/components/details/post-image/index.module.css
@@ -1,0 +1,12 @@
+.container {
+    min-width: 600px;
+    border-left: 1px solid lightgray;
+    border-top: 1px solid lightgray;
+    border-bottom: 1px solid lightgray;
+}
+
+.image {
+    object-fit: cover;
+    max-width: 600px;
+    max-height: 598px;
+}
