+++ src/components/footer/index.module.css
@@ -1,7 +1,5 @@
 .main {
     width: 100%;
-    background-color: white;
-    border-top: 1px solid lightgray;
 }
 
 .footer {
