+++ src/components/logo/index.module.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Pacifico&display=swap');
 
 .logo {
-    font-size: 2.5rem;
+    font-size: 2rem;
     text-align: center;
 }
