+++ src/components/post/likes/index.js
@@ -4,7 +4,9 @@ import styles from './index.module.css'
 const Likes = ({ likes }) => {
     return (
         <div>
-            {likes.length ? <div className={styles.comment}>{likes.length} {likes.length > 1 ? 'likes' : 'like'}</div> : null}
+            {likes.length
+                ? <div className={styles.comment}>{likes.length} {likes.length > 1 ? 'likes' : 'like'}</div>
+                : <div className={styles.noComment}>Be the first to <strong>like this</strong></div>}
         </div>
     )
 }
