+++ src/pages/profile-page/profile-picture/index.module.css
@@ -1,6 +1,6 @@
 .container {
-    width: 150px;
-    height: 150px;
+    max-width: 150px;
+    max-height: 150px;
     margin: 0 5em;
 }
